put addForm into formActions and deleted actions.js + updated import for it
