+++ linkedin-ext/src/actions/formActions.js
@@ -38,3 +38,14 @@ export const DELETE_FORM_FAILURE = "DELETE_FORM_FAILURE";
         alert("Failed to delete form, please try again")
       })
   };
+
+  export const addForm = (newForm) => async dispatch  => {
+    console.log(newForm.name)
+    await axios.post(`https://linkedinextension.herokuapp.com/api/forms/${newForm.id}`, {
+        name: newForm.name,
+    })
+    .then(form => {
+        console.log(form)
+    })
+    .catch(err => console.log(err))
+}
