+++ linkedin-ext/src/components/views/Forms/NewForm.js
@@ -1,6 +1,6 @@
 import React, {Component } from 'react';
 import {connect} from 'react-redux'
-import {addForm} from '../../../actions/actions'
+import {addForm} from '../../../actions/formActions.js'
 import { Redirect } from 'react-router-dom'
 
 import classes from './NewForm.module.css'
